:hammer: fix(ticket) 티켓 트랜잭션 서비스 로직으로 이동#39
