+++ src/database/repositories/ticket.repository.ts
@@ -13,7 +13,7 @@ import { PagingDto } from 'src/common/dtos/paging.dto';
 import { TicketFindDto } from 'src/common/dtos/ticket-find.dto';
 import { UpdateTicketStatusDto } from 'src/common/dtos/update-ticket-status.dto';
 
-import { DataSource, Repository } from 'typeorm';
+import { Repository } from 'typeorm';
 import { Ticket } from '../entities/ticket.entity';
 import { User } from '../entities/user.entity';
 
@@ -21,8 +21,7 @@ import { User } from '../entities/user.entity';
 export class TicketRepository {
   constructor(
     @InjectRepository(Ticket)
-    private ticketRepository: Repository<Ticket>,
-    private dataSource: DataSource
+    private ticketRepository: Repository<Ticket>
   ) {}
 
   /**
@@ -135,42 +134,12 @@ export class TicketRepository {
   }
 
   /**
-   * 해당 ticketId를 참조하여 Ticket 엔티티의 status를 변경하고 DB에 저장한다
-   * @param ticketId Ticket의 id
-   * @param status 변경하고자 하려는 상태
-   * @param admin 변경하려는 어드민 정보
+   * 해당 티켓을 저장한다
+   * @param ticket 저장할 티켓
    */
-  async updateStatus(
-    updateTicketStatus: UpdateTicketStatusDto,
-    admin: User
-  ): Promise<Ticket> {
-    const { ticketId, status } = updateTicketStatus;
-    const queryRunner = this.dataSource.createQueryRunner();
-
-    const ticket = await this.ticketRepository.findOne({
-      where: {
-        id: ticketId
-      }
-    });
-
-    if (!ticket) {
-      throw new NotFoundException(`Can't find Tickets with id: ${ticketId}`);
-    }
-
-    await queryRunner.connect();
-    await queryRunner.startTransaction();
-    try {
-      ticket.status = status;
-      ticket.admin = admin;
-      await queryRunner.commitTransaction();
-    } catch (error) {
-      console.log(`Error occurs in updateStatus: ${error}`);
-      await queryRunner.rollbackTransaction();
-    } finally {
-      await queryRunner.release();
-    }
-
-    return ticket;
+  async saveTicket(ticket: Ticket): Promise<Ticket> {
+    const savedTicket = await this.ticketRepository.save(ticket);
+    return savedTicket;
   }
 
   /**
